Update ClickMoveItem.js
Changed the reference point position calculation, which was previously performed repeatedly each time an Item Object was selected and a command was executed, to not be repeated. When first executed, the coordinate information of the calculated reference point is saved as a Tags Object, and the coordinate information is read each time the command is executed.